SPY-179: Reconnect has wrong ceiling, leading to very long reconnect delays.
Motivation
----------
This is a regression which was introduced through the change in SPY-158
(https://github.com/couchbase/spymemcached/commit/bfaec2b4f1385f20f32fd0ed0736b52ef572e340).

Because the time is now converted to millis right away, but the comparison checks with the 2^n
still converted to seconds. That way, even with a very high N it is smaller than the default of
30k.

Modifications
-------------
The code now correctly converts the 2^n to milliseconds before comparing it, making it a comparison
on the same unit of time again.

Result
------
The ceiling is now properly respected and as a result, reconnect delays are properly capped at
their defined maximum.

Change-Id: I0c976344d0f8e09bc041d94f3544620ef4fb3117
Reviewed-on: http://review.couchbase.org/42925
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>