Remove spymemcached-it targets so it builds
(it = Integration Test)
The file necessary to build this target was not committed upstream,
making it impossible to build.
We also don't need it, so just removing.

Also hardcodes the output name to memcache-asg-java-client
(not AmazonElastiCacheClusterClient via ${name})
to match datomic's artifactId.

Also adds some missing exclusions for `ant clean`.